スラッシュコマンドを移行
