+++ src/bot/commands.ts
@@ -9,27 +9,21 @@ import {
   VoiceBasedChannel,
 } from 'discord.js';
 import ytpl from 'ytpl';
-import { checkUserType, getIntArray, getRndNumber, isEnableFunction } from '../common/common.js';
+import { checkUserType, isEnableFunction } from '../common/common.js';
 import { Logger } from '../common/logger.js';
-import { CONFIG, LiteLLMModel, LITELLM_MODEL } from '../config/config.js';
-import { LiteLLMMode, Role } from '../constant/chat/chat.js';
+import { LITELLM_MODEL } from '../config/config.js';
+import { LiteLLMMode } from '../constant/chat/chat.js';
 import { HELP_COMMANDS, functionNames } from '../constant/constants.js';
-import { ITO_TOPICS } from '../constant/words/ito.js';
-import { TOPIC } from '../constant/words/topic.js';
 import { RoleType } from '../model/models/role.js';
-import { Users, UsersType } from '../model/models/users.js';
+import { UsersType } from '../model/models/users.js';
 import { ColorRepository } from '../model/repository/colorRepository.js';
 import { LogRepository } from '../model/repository/logRepository.js';
 import { PlaylistRepository } from '../model/repository/playlistRepository.js';
 import { RoleRepository } from '../model/repository/roleRepository.js';
 import { UsersRepository } from '../model/repository/usersRepository.js';
-import * as ChatService from '../service/chat.service.js';
 import { LogLevel } from '../type/types.js';
 import * as DotBotFunctions from './dot_function/index.js';
 import { getDefaultRoomName } from './dot_function/voice.js';
-import * as BotFunctions from './function/index.js';
-
-let ITO_NUMS = getIntArray(100);
 
 /**
  * 渡されたコマンドから処理を実行する
@@ -1023,483 +1017,6 @@ export async function commandSelector(message: Message) {
         return;
       }
       throw new Error('再起動');
-    }
-  }
-}
-
-/**
- * 渡されたスラッシュコマンドを処理する
- * @param interaction
- * @deprecated interaction.managerが実装完了後移行する
- * @returns
- */
-export async function interactionSelector(interaction: ChatInputCommandInteraction<CacheType>): Promise<void> {
-  const { commandName } = interaction;
-
-  switch (commandName) {
-    case 'ping': {
-      await interaction.reply('Pong!');
-      break;
-    }
-    case 'help': {
-      await help(interaction);
-      break;
-    }
-    case 'debug': {
-      const url = interaction.options.getString('url');
-      await Logger.put({
-        guild_id: interaction.guild?.id,
-        channel_id: interaction.channel?.id,
-        user_id: interaction.user.id,
-        level: LogLevel.INFO,
-        event: 'received-command/debug',
-        message: [`${url}`],
-      });
-      await interaction.reply('test.');
-      break;
-    }
-    case 'gacha': {
-      await Logger.put({
-        guild_id: interaction.guild?.id,
-        channel_id: interaction.channel?.id,
-        user_id: interaction.user.id,
-        level: LogLevel.INFO,
-        event: 'received-command/gacha',
-        message: [`${interaction}`],
-      });
-      const type = interaction.options.getSubcommand();
-      switch (type) {
-        case 'pick': {
-          await interaction.deferReply();
-          const num = interaction.options.getNumber('num') ?? undefined;
-          const limit = interaction.options.getBoolean('limit') ?? undefined;
-
-          await BotFunctions.Gacha.pickGacha(interaction, limit, num);
-          break;
-        }
-        case 'list': {
-          await BotFunctions.Gacha.getGachaInfo(interaction);
-          break;
-        }
-        case 'extra': {
-          await interaction.deferReply();
-          await BotFunctions.Gacha.extraPick(
-            interaction,
-            interaction.options.getNumber('num') ?? undefined,
-            interaction.options.getString('item') ?? undefined
-          );
-          break;
-        }
-      }
-      break;
-    }
-    case 'gc': {
-      await interaction.deferReply();
-      const num = interaction.options.getNumber('num') ?? undefined;
-      const limit = interaction.options.getBoolean('limit') ?? undefined;
-
-      await BotFunctions.Gacha.pickGacha(interaction, limit, num);
-      break;
-    }
-    case 'gl': {
-      await interaction.deferReply();
-      await BotFunctions.Gacha.pickGacha(interaction, true);
-      break;
-    }
-    case 'dice': {
-      await interaction.deferReply({ ephemeral: true });
-      const num = interaction.options.getNumber('num') ?? 1;
-      const max = interaction.options.getNumber('max') ?? 100;
-      await BotFunctions.Dice.rollHideDice(interaction, num, max);
-      break;
-    }
-    case 'genito': {
-      await interaction.deferReply();
-      ITO_NUMS = getIntArray(100);
-      const nums = [];
-      for (let i = 0; i < 4; i++) {
-        const num = getRndNumber(0, ITO_TOPICS.length - 1);
-        nums.push(num);
-      }
-      const send = new EmbedBuilder()
-        .setColor('#ff9900')
-        .setTitle(`今回のお題は？`)
-        .setDescription(nums.map((num) => ITO_TOPICS[num]).join('\n'));
-
-      interaction.editReply({ embeds: [send] });
-      break;
-    }
-    case 'ito': {
-      await interaction.deferReply({ ephemeral: true });
-      const round = interaction.options.getNumber('round') ?? 1;
-      // roundの数だけ取り出す
-      const sendNums = ITO_NUMS.slice(0, round);
-      ITO_NUMS.splice(0, round);
-
-      await Logger.put({
-        guild_id: interaction.guild?.id,
-        channel_id: interaction.channel?.id,
-        user_id: interaction.user.id,
-        level: LogLevel.INFO,
-        event: 'received-command/ito',
-        message: [`${sendNums.join(', ')}`],
-      });
-
-      interaction.editReply({ content: `貴方の数は${sendNums.join(', ')}です.` });
-      break;
-    }
-    case 'memory': {
-      await BotFunctions.Chat.setMemory(interaction);
-      break;
-    }
-    case 'topic': {
-      const num = getRndNumber(0, TOPIC.length - 1);
-      const send = new EmbedBuilder().setColor('#ff9900').setTitle(`こんなのでました～！`).setDescription(TOPIC[num]);
-
-      interaction.reply({ embeds: [send] });
-      break;
-    }
-    case 'dc': {
-      // required option is not null
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      const user = interaction.options.getUser('user')!;
-      if (!interaction.guild) {
-        return;
-      }
-      if (!checkUserType(interaction.guild.id, interaction.user.id, UsersType.ADMIN)) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`このコマンドは管理者のみ使用できます。`);
-
-        interaction.reply({ embeds: [send] });
-        return;
-      }
-      const id = user.id;
-      if (!id) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`id not found or invalid`);
-
-        interaction.reply({ embeds: [send] });
-        return;
-      }
-
-      if (interaction.channel?.type === ChannelType.GuildVoice) {
-        const channel = interaction.channel as BaseGuildVoiceChannel;
-        const member = channel.members.find((member) => member.id === id);
-        if (!member) {
-          const send = new EmbedBuilder()
-            .setColor('#ff0000')
-            .setTitle(`エラー`)
-            .setDescription(`id not found or invalid`);
-
-          interaction.reply({ embeds: [send] });
-          break;
-        }
-        await member.voice.disconnect();
-
-        await Logger.put({
-          guild_id: interaction.guild?.id,
-          channel_id: interaction.channel?.id,
-          user_id: interaction.user.id,
-          level: LogLevel.INFO,
-          event: 'disconnect-user',
-          message: [`disconnect ${member.user.displayName} by ${interaction.user}`],
-        });
-
-        const send = new EmbedBuilder()
-          .setColor('#00ff00')
-          .setTitle(`成功`)
-          .setDescription(`${member.user.displayName}を切断しました`);
-
-        interaction.reply({ embeds: [send] });
-      }
-      break;
-    }
-    case 'timeout': {
-      const guild = interaction.guild;
-      if (!guild) {
-        return;
-      }
-      if (!checkUserType(interaction.guild.id, interaction.user.id, UsersType.ADMIN)) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`このコマンドは管理者のみ使用できます。`);
-
-        interaction.reply({ embeds: [send] });
-        return;
-      }
-
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      const user = interaction.options.getUser('user')!;
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      const time = interaction.options.getNumber('time')!;
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      const reason = interaction.options.getString('reason')!;
-      const member = guild.members.cache.find((member) => member.id === user.id);
-
-      if (!member) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription('ユーザーが見つからなかった');
-        await interaction.reply({ embeds: [send] });
-        return;
-      }
-
-      let dmChannel = user.dmChannel;
-      if (!dmChannel) {
-        // 一度もDMを送信していない場合
-        dmChannel = await user.createDM();
-      }
-
-      const message = await ChatService.getTimeoutMessage(interaction.user.id, user.id, reason, time);
-
-      const dmSend = new EmbedBuilder()
-        .setColor('#ff0000')
-        .setTitle(`⚠️警告：タイムアウト処分が行われました`)
-        .setDescription(message);
-      await dmChannel.send({ embeds: [dmSend] });
-
-      await member.timeout(time * 60 * 60 * 1000, reason);
-
-      const send = new EmbedBuilder()
-        .setColor('#00ff00')
-        .setTitle(`タイムアウトの実行`)
-        .setFields(
-          { name: '対象人物', value: member.user.displayName },
-          { name: '規制時間', value: `${time.toString()} 時間` },
-          { name: '理由', value: reason }
-        );
-      await interaction.reply({ embeds: [send] });
-
-      const chatService = new ChatService.InteractionChatService(interaction);
-      await chatService.addChat(Role.USER, message);
-      return;
-    }
-    case 'mute': {
-      const guild = interaction.guild;
-      if (!guild) {
-        return;
-      }
-      if (!checkUserType(interaction.guild.id, interaction.user.id, UsersType.ADMIN)) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription(`このコマンドは管理者のみ使用できます。`);
-
-        interaction.reply({ embeds: [send] });
-        return;
-      }
-
-      const user = interaction.options.getUser('user')!;
-      const time = interaction.options.getNumber('time')!;
-      const reason = interaction.options.getString('reason')!;
-      const member = guild.members.cache.find((member) => member.id === user.id);
-
-      if (!member) {
-        const send = new EmbedBuilder()
-          .setColor('#ff0000')
-          .setTitle(`エラー`)
-          .setDescription('ユーザーが見つからなかった');
-        await interaction.reply({ embeds: [send] });
-        return;
-      }
-
-      member.voice.setMute(true, reason);
-
-      let dmChannel = user.dmChannel;
-      if (!dmChannel) {
-        // 一度もDMを送信していない場合
-        dmChannel = await user.createDM();
-      }
-
-      const message = await ChatService.getMuteMessage(interaction.user.id, user.id, reason, time);
-
-      const dmSend = new EmbedBuilder()
-        .setColor('#ff0000')
-        .setTitle(`⚠️警告：マイクミュート処分が行われました`)
-        .setDescription(message);
-      await dmChannel.send({ embeds: [dmSend] });
-
-      const send = new EmbedBuilder()
-        .setColor('#00ff00')
-        .setTitle(`ミュートの実行`)
-        .setDescription(`${member.user.displayName}をサーバーミュートしました.`);
-      send.addFields({ name: '規制時間', value: `${time.toString()} 分` });
-      send.addFields({ name: '事由', value: reason });
-      await interaction.reply({ embeds: [send] });
-
-      const chatService = new ChatService.InteractionChatService(interaction);
-      await chatService.addChat(Role.USER, message);
-
-      setTimeout(
-        async () => {
-          const dmSend = new EmbedBuilder()
-            .setColor('#00ff00')
-            .setTitle(`マイクミュートの解除`)
-            .setDescription(`${member.user.displayName}のマイクミュートを解除しました.`);
-          await dmChannel.send({ embeds: [dmSend] });
-          member.voice.setMute(false);
-        },
-        time * 60 * 1000
-      );
-
-      break;
-    }
-    case 'accept': {
-      if (!interaction.guild) {
-        return;
-      }
-      if (interaction.channel?.type === ChannelType.GuildText) {
-        const user = interaction.user;
-        await Logger.put({
-          guild_id: interaction.guild.id,
-          channel_id: interaction.channel.id,
-          user_id: user.id,
-          level: LogLevel.INFO,
-          event: 'reaction-add',
-          message: [`rule accepted: ${user.displayName}`],
-        });
-
-        const u = interaction.guild.members.cache.get(user.id);
-
-        if (!u) {
-          console.error('user not found');
-          return;
-        }
-
-        const roleRepository = new RoleRepository();
-
-        const r = await roleRepository.getRoleByName(interaction.guild.id, 'member');
-        if (!r) {
-          console.error('role not found');
-          return;
-        }
-        const userRole = u?.roles.cache.find((role) => role.id === r.role_id);
-
-        await Logger.put({
-          guild_id: interaction.guild.id,
-          channel_id: interaction.channel.id,
-          user_id: user.id,
-          level: LogLevel.INFO,
-          event: 'role-check',
-          message: [u?.roles.cache.map((role) => role.name).join(',')],
-        });
-        if (userRole) {
-          const message = await interaction.reply({ content: `もうロールが付いてるみたい！`, ephemeral: true });
-          setTimeout(async () => {
-            await message.delete();
-          }, 3000);
-          return;
-        }
-
-        // add user role
-        await u?.roles.add(r.role_id);
-
-        // register user
-        const userRepository = new UsersRepository();
-        const userEntity = await userRepository.get(interaction.guild.id, user.id);
-        if (!userEntity) {
-          const saveUser: Partial<Users> = {
-            id: user.id,
-            user_name: user.displayName,
-            pick_left: 10,
-            voice_channel_data: [
-              {
-                gid: interaction.guild.id ?? 'DM',
-                date: new Date(),
-              },
-            ],
-          };
-          await userRepository.save(saveUser);
-        }
-        const name = u?.roles.cache.find((role) => role.name === 'member')?.name;
-        await Logger.put({
-          guild_id: interaction.guild.id,
-          channel_id: interaction.channel.id,
-          user_id: user.id,
-          level: LogLevel.INFO,
-          event: 'add-role',
-          message: name ? [name] : undefined,
-        });
-
-        const message = await interaction.reply({
-          content: `読んでくれてありがと～！ロールを付与したよ！`,
-          ephemeral: true,
-        });
-        setTimeout(async () => {
-          await message.delete();
-        }, 3000);
-        return;
-      }
-      break;
-    }
-    case 'room': {
-      const type = interaction.options.getSubcommand();
-      await interaction.deferReply({ ephemeral: true });
-      switch (type) {
-        case 'create': {
-          const name = interaction.options.getString('name');
-          const isLive = interaction.options.getBoolean('live') ?? false;
-          const isPrivate = interaction.options.getBoolean('private') ?? true;
-
-          if (!name) {
-            await interaction.editReply({ content: 'チャンネル名を指定してください' });
-            return;
-          }
-          await BotFunctions.Room.createRoom(interaction, name, isLive, isPrivate);
-          break;
-        }
-        case 'add': {
-          const user = interaction.options.getUser('user');
-          const member = interaction.guild?.members.cache.get(user?.id ?? '');
-          if (!member) {
-            await interaction.editReply({ content: 'ユーザーが見つかりません' });
-            return;
-          }
-          await BotFunctions.Room.addPermission(interaction, member);
-          break;
-        }
-        case 'remove': {
-          const user = interaction.options.getUser('user');
-          const member = interaction.guild?.members.cache.get(user?.id ?? '');
-          if (!member) {
-            await interaction.editReply({ content: 'ユーザーが見つかりません' });
-            return;
-          }
-          await BotFunctions.Room.removePermission(interaction, member);
-          break;
-        }
-        case 'lock': {
-          await BotFunctions.Room.toggleAutoDelete(interaction);
-          break;
-        }
-      }
-      break;
-    }
-    case 'nickname': {
-      const name = interaction.options.getString('name');
-      if (!name) {
-        return;
-      }
-      if (interaction.guild) {
-        const usersRepository = new UsersRepository();
-        const user = await usersRepository.get(interaction.guild.id, interaction.user.id);
-        if (!user) {
-          await interaction.reply({ content: `あなたのことが登録されていないみたい……？` });
-          return;
-        }
-        await usersRepository.save({ ...user, userSetting: { ...user.userSetting, nickname: name } });
-        await interaction.reply({ content: `はーい！これから「${name}」って呼ぶね～！` });
-      }
-      break;
-    }
-    default: {
-      return;
     }
   }
 }
