+++ src/bot/manager/handlers/interactions/memory.handler.ts
@@ -1,0 +1,14 @@
+import { CacheType, ChatInputCommandInteraction } from 'discord.js';
+import { BaseInteractionHandler } from '../../interaction.handler.js';
+import { Logger } from '../../../../common/logger.js';
+import * as BotFunctions from '../../../function/index.js';
+
+export class MemoryHandler extends BaseInteractionHandler {
+  constructor(logger?: Logger) {
+    super(logger);
+  }
+
+  async execute(interaction: ChatInputCommandInteraction<CacheType>): Promise<void> {
+    await BotFunctions.Chat.setMemory(interaction);
+  }
+}
